move scheduler.py to utils
